+++ src/RateLimiter.ts
@@ -1,0 +1,27 @@
+import type { ValueFilter } from "./FilteredObservable";
+
+export enum RateLimitType {
+	Debounce,
+	Throttle
+}
+
+export function RateLimiter<T>(type: RateLimitType, delay: number): ValueFilter<T> {
+	let timeoutId: NodeJS.Timeout|undefined = undefined;
+	let pendingValue: T;
+
+	return (newValue: T, setValue: (newValue: T) => void): void => {
+		pendingValue = newValue;
+
+		if (type === RateLimitType.Debounce && timeoutId !== undefined) {
+			clearTimeout(timeoutId);
+			timeoutId = undefined;
+		}
+
+		if (timeoutId === undefined) {
+			timeoutId = setTimeout(() => {
+				timeoutId = undefined;
+				setValue(pendingValue);
+			}, delay);
+		}
+	};
+}