+++ tests/Observable.test.ts
@@ -14,33 +14,27 @@ beforeEach(() => {
 test("Should have value", () => {
 	const t = new Observable(true);
 
-	expect(t.Value).toBe(true);
+	expect(t.Value).toStrictEqual(true);
 });
 
 test("Should allow value to change", () => {
 	const t = new Observable(true);
 
-	expect(t.Value).toBe(true);
+	expect(t.Value).toStrictEqual(true);
 
 	t.Value = false;
 
-	expect(t.Value).toBe(false);
+	expect(t.Value).toStrictEqual(false);
 });
 
 test("Should allow value to be transformed", () => {
 	const t = new Observable(true);
 
-	expect(t.Value).toBe(true);
+	expect(t.Value).toStrictEqual(true);
 
 	t.Update(x => !x);
 
-	expect(t.Value).toBe(false);
-});
-
-test("Should be usable as a read only observable", () => {
-	const t = new Observable(true);
-
-	expect(t.AsReadOnly.Value).toBe(true);
+	expect(t.Value).toStrictEqual(false);
 });
 
 test("Should notify observers on change until they unsubscribe", () => {
@@ -48,13 +42,13 @@ test("Should notify observers on change until they unsubscribe", () => {
 	const unsubscribe = t.Subscribe(mockObserver);
 
 	t.Value = false;
-	ThenObserverWasCalled(mockObserver, 1, false, true);
+	ThenObserverWasCalled(mockObserver, 1, false);
 
 	unsubscribe();
 
 	t.Value = true;
 
-	expect(mockObserver.mock.calls.length).toBe(1);
+	ThenObserverCallCountIs(mockObserver, 1);
 });
 
 test("Should notify all observers on change until they unsubscribe", () => {
@@ -63,14 +57,14 @@ test("Should notify all observers on change until they unsubscribe", () => {
 	const unsubscribeTwo = t.Subscribe(mockObserverTwo);
 
 	t.Value = false;
-	ThenObserverWasCalled(mockObserver, 1, false, true);
-	ThenObserverWasCalled(mockObserverTwo, 1, false, true);
+	ThenObserverWasCalled(mockObserver, 1, false);
+	ThenObserverWasCalled(mockObserverTwo, 1, false);
 
 	unsubscribeTwo();
 
 	t.Value = true;
 
-	ThenObserverWasCalled(mockObserver, 2, true, false);
+	ThenObserverWasCalled(mockObserver, 2, true);
 	ThenObserverCallCountIs(mockObserverTwo, 1);
 
 	unsubscribe();
@@ -86,7 +80,7 @@ test("Should notify all observers on change until they unsubscribe", () => {
 
 	ThenObserverCallCountIs(mockObserver, 2);
 	ThenObserverCallCountIs(mockObserverTwo, 1);
-	ThenObserverWasCalled(mockObserverThree, 1, true, false);
+	ThenObserverWasCalled(mockObserverThree, 1, true);
 
 	unsubscribeThree();
 
@@ -111,11 +105,11 @@ test("Should deal with observer being removed multiple times", () => {
 
 	const unsubscribe = t.Subscribe(mockObserver);
 
-	expect(t.Value).toBe(true);
+	expect(t.Value).toStrictEqual(true);
 
 	t.Value = false;
 
-	ThenObserverWasCalled(mockObserver, 1, false, true);
+	ThenObserverWasCalled(mockObserver, 1, false);
 
 	unsubscribe();
 	unsubscribe();