+++ tests/ReadOnlyObservable.test.ts
@@ -1,7 +1,0 @@
-import { ReadOnlyObservable } from "../src/ReadOnlyObservable";
-
-test("Should have value", () => {
-	const t = new ReadOnlyObservable(true);
-
-	expect(t.Value).toBe(true);
-});